requirements.txt
discord.py
ossapi == 5.2.0
requests
pytz
pymongo[srv] >= 4.11.3
boto3